+++ akshare/bond/bond_china_money.py
@@ -1,23 +1,45 @@
 #!/usr/bin/env python
 # -*- coding:utf-8 -*-
 """
-Date: 2022/2/2 22:43
+Date: 2022/12/6 15:43
 Desc: 收盘收益率曲线历史数据
-http://www.chinamoney.com.cn/chinese/bkcurvclosedyhis/?bondType=CYCC000&reference=1
+https://www.chinamoney.com.cn/chinese/bkcurvclosedyhis/?bondType=CYCC000&reference=1
 """
 import pandas as pd
 import requests
+from functools import lru_cache
 
 
+@lru_cache()
 def bond_china_close_return_map() -> pd.DataFrame:
     """
     收盘收益率曲线历史数据
-    http://www.chinamoney.com.cn/chinese/bkcurvclosedyhis/?bondType=CYCC000&reference=1
+    https://www.chinamoney.com.cn/chinese/bkcurvclosedyhis/?bondType=CYCC000&reference=1
     :return: 收盘收益率曲线历史数据
     :rtype: pandas.DataFrame
     """
     url = "http://www.chinamoney.com.cn/ags/ms/cm-u-bk-currency/ClsYldCurvCurvGO"
-    r = requests.post(url)
+    headers = {
+        "Accept": "application/json, text/javascript, */*; q=0.01",
+        "Accept-Encoding": "gzip, deflate, br",
+        "Accept-Language": "zh-CN,zh;q=0.9,en;q=0.8",
+        "Cache-Control": "no-cache",
+        "Connection": "keep-alive",
+        "Content-Length": "0",
+        "Host": "www.chinamoney.com.cn",
+        "Origin": "https://www.chinamoney.com.cn",
+        "Pragma": "no-cache",
+        "Referer": "https://www.chinamoney.com.cn/chinese/bkcurvclosedyhis/?bondType=CYCC000&reference=1",
+        "sec-ch-ua": '"Not?A_Brand";v="8", "Chromium";v="108", "Google Chrome";v="108"',
+        "sec-ch-ua-mobile": "?0",
+        "sec-ch-ua-platform": '"Windows"',
+        "Sec-Fetch-Dest": "empty",
+        "Sec-Fetch-Mode": "cors",
+        "Sec-Fetch-Site": "same-origin",
+        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36",
+        "X-Requested-With": "XMLHttpRequest"
+    }
+    r = requests.post(url, headers=headers)
     data_json = r.json()
     temp_df = pd.DataFrame(data_json["records"])
     return temp_df
@@ -26,12 +48,12 @@ def bond_china_close_return_map() -> pd.DataFrame:
 def bond_china_close_return(
     symbol: str = "国债",
     period: str = "1",
-    start_date: str = "20220311",
-    end_date: str = "20220311",
+    start_date: str = "20221111",
+    end_date: str = "20221211",
 ) -> pd.DataFrame:
     """
     收盘收益率曲线历史数据
-    http://www.chinamoney.com.cn/chinese/bkcurvclosedyhis/?bondType=CYCC000&reference=1
+    https://www.chinamoney.com.cn/chinese/bkcurvclosedyhis/?bondType=CYCC000&reference=1
     :param symbol: 需要获取的指标
     :type period: choice of {'0.1', '0.5', '1'}
     :param period: 期限间隔
@@ -46,6 +68,9 @@ def bond_china_close_return(
     name_code_df = bond_china_close_return_map()
     symbol_code = name_code_df[name_code_df["cnLabel"] == symbol]["value"].values[0]
     url = "https://www.chinamoney.com.cn/ags/ms/cm-u-bk-currency/ClsYldCurvHis"
+    headers = {
+        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/108.0.0.0 Safari/537.36",
+    }
     params = {
         "lang": "CN",
         "reference": "1,2,3",
@@ -56,9 +81,7 @@ def bond_china_close_return(
         "pageNum": "1",
         "pageSize": "15",
     }
-    headers = {
-        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/96.0.4664.110 Safari/537.36'
-    }
+
     r = requests.get(url, params=params, headers=headers)
     data_json = r.json()
     temp_df = pd.DataFrame(data_json["records"])
@@ -89,9 +112,9 @@ def bond_china_close_return(
 
 if __name__ == "__main__":
     bond_china_close_return_df = bond_china_close_return(
-        symbol="国债", period="1", start_date="20220104", end_date="20220104"
+        symbol="国债", period="1", start_date="20221111", end_date="20221211"
     )
     print(bond_china_close_return_df)
 
-    bond_china_close_return_df = bond_china_close_return(symbol="政策性金融债(进出口行)", period="1", start_date="20220104", end_date="20220104")
+    bond_china_close_return_df = bond_china_close_return(symbol="政策性金融债(进出口行)", period="1", start_date="20221111", end_date="20221211")
     print(bond_china_close_return_df)