+++ docs/changelog.md
@@ -43,6 +43,10 @@
 
 ## 更新说明
 
+1.8.26 fix: fix bond_china_close_return_map interface
+
+    1. 修复 bond_china_close_return_map 接口
+
 1.8.25 fix: fix stock_circulate_stock_holder interface
 
     1. 修复 stock_circulate_stock_holder 接口
@@ -1634,6 +1638,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.8.26 fix: fix bond_china_close_return_map interface
 
 1.8.25 fix: fix stock_circulate_stock_holder interface